Move report generation functionality to task
